LKLN-45 Feat:Raffle-API-Controller 내용: Item Merge전 최신 커밋
